Create configObj interface
