+++ packages/optimizely-sdk/lib/core/project_config/index.d.ts
@@ -13,38 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { LogHandler, ErrorHandler } from '@optimizely/js-sdk-logging';
-import { DatafileOptions } from '@optimizely/optimizely-sdk';
-import { EventDispatcher } from '@optimizely/js-sdk-event-processor';
-import { UserProfileService } from '../../shared_types';
+import { LogHandler } from '@optimizely/js-sdk-logging';
+import { DatafileOptions } from '../../shared_types';
 import { FeatureFlag, FeatureVariable, Experiment, Variation  } from './entities';
 
 export interface ProjectConfig {
-  revision: string;
-  projectId: string;
-  featureKeyMap?: {
-    [key: string]: FeatureFlag
-  };
-  clientEngine: string;
-  clientVersion?: string;
-  errorHandler: ErrorHandler;
-  eventDispatcher: EventDispatcher;
-  isValidInstance: boolean;
+  revision?: string;
+  projectId?: string;
   // TODO[OASIS-6649]: Don't use object type
   // eslint-disable-next-line  @typescript-eslint/ban-types
   datafile: object | string;
+  datafileOptions?: DatafileOptions;
   // TODO[OASIS-6649]: Don't use object type
   // eslint-disable-next-line  @typescript-eslint/ban-types
   jsonSchemaValidator?: object;
   sdkKey?: string;
-  userProfileService?: UserProfileService | null;
-  UNSTABLE_conditionEvaluators?: unknown;
-  eventFlushInterval?: number;
-  eventBatchSize?: number;
-  datafileOptions?: DatafileOptions;
-  eventMaxQueueSize?: number;
-  logger: LogHandler;
-  experimentKeyMap:{[key: string]: Experiment};
+  experimentKeyMap?:{[key: string]: Experiment};
+  featureKeyMap?: {
+    [key: string]: FeatureFlag
+  };
 }
 /**
  * Determine for given experiment if event is running, which determines whether should be dispatched or not
